Clean up auctionhouse API and refine test cases
