+++ test/utils.ts
@@ -15,6 +15,7 @@ export const THOUSANDTH_ETH = ethers.utils.parseUnits(
   "0.001",
   "ether"
 ) as BigNumber;
+export const HUNDRETH_ETH = ethers.utils.parseUnits("0.01", "ether") as BigNumber;
 export const TENTH_ETH = ethers.utils.parseUnits("0.1", "ether") as BigNumber;
 export const ONE_ETH = ethers.utils.parseUnits("1", "ether") as BigNumber;
 export const TWO_ETH = ethers.utils.parseUnits("2", "ether") as BigNumber;
